requirements.txt
requests>=2.28.1
python-nmap>=0.7.1
colorama>=0.4.5
tabulate>=0.9.0
jsonschema>=4.17.3
tqdm>=4.64.1